Arguments extended, compare to any pair possible

+++ standardfollowup/src/main/java/standard/CollectAndEvaluate.java
@@ -2,6 +2,7 @@ package standard;
 
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -10,71 +11,109 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
+import java.util.concurrent.Callable;
 import java.util.stream.Collectors;
 
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 
+import picocli.CommandLine;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
 import standard.ec.publichealth.ReadPublicHealth;
 
 /**
  * Collect documents of medical regulation
  * Compare with last verions and report differences
  */
-public class CollectAndEvaluate {
+@Command(name = "checksum", mixinStandardHelpOptions = true, version = "checksum 4.0", description = "Prints the checksum (SHA-256 by default) of a file to STDOUT.")
+public class CollectAndEvaluate implements Callable<Integer> {
 
 	private static final String PROTOCOLL_TXT = "protocoll.txt";
-	private final CommonFunctions common;
-	
+	private static final String COMPARATOR_CMD = "compare.cmd";
+	private CommonFunctions common;
+	private Path oldDir;
+	private Path newDir;
+
+	@Option(names = { "-b", "--base" }, defaultValue="C:/temp/Standards", description = "Base of all collected standardsthe archive file")
+	Path baseDir;
+
+	@Option(names = { "-o", "--old" }, description = "Directory to compare with")
+	String oldDirArg;
+
+	@Option(names = { "-n", "--new" }, description = "Directory to compare with")
+	String newDirArg;
+
 	/**
 	 * @param Directory of collected documents or c:\temp\Standards if absent
 	 */
 	public static void main(String[] args) throws IOException, URISyntaxException {
-		new CollectAndEvaluate(args);
+		int exitCode = new CommandLine(new CollectAndEvaluate()).execute(args);
+		System.exit(exitCode);
 	}
-	
-	public CollectAndEvaluate(String[] args) throws IOException, URISyntaxException {
-		common = new CommonFunctions("PublicHealth", "https://health.ec.europa.eu", args);
-		collect();
+
+	@Override
+	public Integer call() throws Exception {
+		doIt();
+		return 0;
+	}
+
+	public CollectAndEvaluate() throws IOException, URISyntaxException {
+	}
+
+	private void doIt() throws IOException, URISyntaxException {
+		common = new CommonFunctions("PublicHealth", "https://health.ec.europa.eu", baseDir);
+		oldDir = oldDirArg == null ? null : common.pool.resolve(oldDirArg);
+		if (oldDir != null && !Files.isDirectory(oldDir)) {
+			System.out.println("Not correct directory: " + oldDirArg);
+			throw new FileNotFoundException(oldDirArg);
+		}
+		newDir = newDirArg == null ? null : common.pool.resolve(newDirArg);
+		if (newDir != null && !Files.isDirectory(newDir)) {
+			System.out.println("Not correct directory: " + newDirArg);
+			throw new FileNotFoundException(oldDirArg);
+		}
+
+		if (newDirArg == null) {
+			common.createNewDir();
+			collect();
+		}
 		evaluate();
 	}
-	
+
 	private void collect() throws IOException, URISyntaxException {
 		loadEulex();
 		loadHealthcare();
 		loadEVS();
-		
+
 	}
-	
+
 	private void loadEulex() throws IOException, URISyntaxException {
-//		common.downloadFile("https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32017R0745");
+		//		common.downloadFile("https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32017R0745");
 		URI uri = new URI("https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32017R0745");
 		Path targetFile = common.getNewFilesDir().resolve("EU-Medizinprodukte.pdf");
 		FileUtils.copyURLToFile(uri.toURL(), targetFile.toFile());
 	}
-	
+
 	private void loadHealthcare() throws IOException, URISyntaxException {
 		ReadPublicHealth obj = new ReadPublicHealth();
 		obj.downloadFiles();
 	}
-	
+
 	private void loadEVS() {
-		
+
 	}
-	
+
 	private void evaluate() throws IOException {
-		compareFiles();
+		Path oldFilesDir = oldDir == null ? common.getLastButOneDir() : oldDir;
+		Path newFilesDir = newDir == null ? common.newFilesDir : newDir;
+		compareFiles(oldFilesDir, newFilesDir);
 	}
-	
-	private void compareFiles() throws IOException {
-		List<File> subdirs = new ArrayList<>(Arrays.asList(common.getPool().toFile().listFiles()));
-		Collections.sort(subdirs);
-		Path oldFilesDir = subdirs.get(subdirs.size()-2).toPath();
-		Path newFilesDir = subdirs.get(subdirs.size()-1).toPath();
+
+	private void compareFiles(Path oldFilesDir, Path newFilesDir) throws IOException {
 
 		List<String> newFiles = Arrays.asList(newFilesDir.toFile().listFiles()).stream().map(x -> x.getName()).collect(Collectors.toList());
 		List<String> oldFiles = Arrays.asList(oldFilesDir.toFile().listFiles()).stream().map(x -> x.getName()).collect(Collectors.toList());
@@ -89,41 +128,51 @@ public class CollectAndEvaluate {
 		droppedSince.remove(PROTOCOLL_TXT);
 		Collection<String> keptSince = CollectionUtils.intersection(newFiles, oldFiles);
 		List<String> differentFiles = new ArrayList<>();
-		try (BufferedWriter writer =  Files.newBufferedWriter(newFilesDir.resolve(PROTOCOLL_TXT))) {
-			for (String keptFile : keptSince) {
-				Path oldFilePath = oldFilesDir.resolve(keptFile);
-				File oldFile = oldFilePath.toFile();
-				File newFile = newFilesDir.resolve(keptFile).toFile();
-				if (common.areFilesIdentical(oldFile, newFile)) {
-					Files.delete(oldFilePath);
-				} else {
-					differentFiles.add(keptFile);
-					writer.append("cmp ");
-					writer.append(oldFile.toString());
-					writer.append(" ");
-					writer.append(newFile.toString());
+		try (BufferedWriter logger =  Files.newBufferedWriter(newFilesDir.resolve(PROTOCOLL_TXT))) {
+			try (BufferedWriter comparator =  Files.newBufferedWriter(newFilesDir.resolve(COMPARATOR_CMD))) {
+				comparator.append("@set comparator=\"c:\\Program Files\\PDF24\\pdf24-Toolbox.exe\" verb=compare ");
+				comparator.newLine();
+				for (String keptFile : keptSince) {
+					Path oldFilePath = oldFilesDir.resolve(keptFile);
+					File oldFile = oldFilePath.toFile();
+					File newFile = newFilesDir.resolve(keptFile).toFile();
+					if (common.areFilesIdentical(oldFile, newFile)) {
+						if (oldDirArg == null && newDirArg == null) {
+							Files.delete(oldFilePath);
+						}
+					} else {
+						differentFiles.add(keptFile);
+						logger.append("cmp ");
+						logger.append(oldFile.toString());
+						logger.append(" ");
+						logger.append(newFile.toString());
+						comparator.append(String.format("%%comparator%% \"%s\" \"%s\"%n", oldFile.toString(), newFile.toString()));
+						comparator.append("pause");
+						comparator.newLine();
+					}
 				}
+				logger.append(String.format("REM Comparing %s with %s%n", newFilesDir, oldFilesDir));
+				printList(droppedSince, "Dropped", "No standard dropped!", logger);
+				printList(createdSince, "New standard", "No new standard!", logger);
+				printList(differentFiles, "Changed", "No changes!", logger);
 			}
-			writer.append(String.format("REM Comparing %s with %s%n", newFilesDir, oldFilesDir));
-			printList(droppedSince, "Dropped", "No standard dropped!", writer);
-			printList(createdSince, "New standard", "No new standard!", writer);
-			printList(differentFiles, "Changed", "No changes!", writer);
 		}
 		System.out.format("Compared %s with %s%n", newFilesDir, oldFilesDir);
 
 	}
-	
-	private void printList(Collection<String> list, String prefix, String emptyText, BufferedWriter writer) throws IOException {
+
+	private void printList(Collection<String> listOfFiles, String prefix, String emptyText, 
+			BufferedWriter logger) throws IOException {
 		int counter = 1;
-		if (list.isEmpty()) {
+		if (listOfFiles.isEmpty()) {
 			System.out.println(emptyText);
-			writer.append(emptyText);
-			writer.newLine();
+			logger.append(emptyText);
+			logger.newLine();
 		} else {
-			for (String name : list) {
+			for (String name : listOfFiles) {
 				String text = String.format("%3d %s: %s%n", counter, prefix, name);
 				System.out.print(text);
-				writer.append(text);
+				logger.append(text);
 				counter++;
 			}
 		}
